Revert "update policy code and test"
This reverts commit 38a21e1969b2c60fa3cfca2b306432a8722bed65.

# Conflicts:
#	TrafficSimulator/src/com/kcl/keepitclean/test/policy/TestPolicy.java

+++ TrafficSimulator/src/com/kcl/keepitclean/main/policy/SpeedLimit.java
@@ -10,11 +10,11 @@ package com.kcl.keepitclean.main.policy;
  * @author Huong
  */
 public class SpeedLimit {
-    protected  int junction;
-    protected  int straightRoad;
-    protected  int curvyRoad;
+    protected double junction;
+    protected double straightRoad;
+    protected double curvyRoad;
 
-    protected  SpeedLimit(int junctionLimit, int straightRoadLimit, int curvyRoadLimit){
+    protected  SpeedLimit(double junctionLimit, double straightRoadLimit, double curvyRoadLimit){
         this.junction = junctionLimit;
         this.straightRoad = straightRoadLimit;
         this.curvyRoad = curvyRoadLimit;
